# Conflicts:
#	frontend/src/main/scala/org/seekloud/thor/front/thorClient/GameHolder4Play.scala

+++ shared/src/main/scala/org/seekloud/thor/shared/ptcl/component/Food.scala
@@ -7,15 +7,15 @@ import org.seekloud.thor.shared.ptcl.model
 /**
   * Created by Jingyi on 2018/11/9
   */
-case class FoodState(fId: Long, level:Byte, position: Point, radius: Float, color: Int)
+case class FoodState(fId: Int, level:Byte, position: Point, radius: Float, color: Byte)
 
 trait Food extends CircleObjectOfGame {
 
-  val fId: Long
+  val fId: Int
 
   var level: Byte
 
-  var color: Int
+  var color: Byte
 
   override protected var position: Point
 
@@ -33,9 +33,9 @@ object Food {
 }
 
 case class AddNormalFood(
-                        fId: Long,
+                        fId: Int,
                         var level: Byte,
                         var position: model.Point,
                         var radius: Float,
-                        var color: Int
+                        var color: Byte
                         ) extends Food
